+++ src/code_800029B0.h
@@ -102,7 +102,7 @@ extern size_t gFreeMemorySize;
 extern uintptr_t gNextFreeMemoryAddress;
 extern uintptr_t gHeapEndPtr;
 
-extern u32 D_8015F730;
+extern uintptr_t D_8015F730;
 extern uintptr_t gFreeMemoryResetAnchor;
 extern Vec3f D_8015F738;
 extern Vec3f D_8015F748;
